+++ laravel/app/Http/Controllers/ProductController.php
@@ -12,6 +12,7 @@ class ProductController extends Controller
     }
 
     public function createProduct(Request $request) {
+        abort_unless(auth()->user()?->can('products.create'), 403, 'You do not have permission to create products.');
         $product = Product::create($request->all());
         return response()->json($product, 201);
     }
@@ -21,12 +22,14 @@ class ProductController extends Controller
     }
 
     public function updateProduct(Request $request, $productId) {
+        abort_unless(auth()->user()?->can('products.update'), 403, 'You do not have permission to update products.');
         $product = Product::findOrFail($productId);
         $product->update($request->all());
         return $product;
     }
 
     public function deleteProduct($productId) {
+        abort_unless(auth()->user()?->can('products.delete'), 403, 'You do not have permission to delete products.');
         $product = Product::findOrFail($productId);
         $product->delete();
         return response()->json(["message" => "Product deleted"]);
